requirement.txt
api==0.0.7
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
django-rest-registration==0.4.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-extensions==0.6.0
drf-yasg==1.20.0
environ==1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nose==1.3.7
packaging==20.7
Pillow==8.0.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2